Move Link/Links type to external linking code
